Move forgot password page store to central auth store
