--- src/containers/ForgotPasswordPage/constants.js
@@ -1,5 +1,0 @@
-export const FORGOT_PASSWORD = '[Forgot Password]';
-
-export const FORGOT_PASSWORD_REQUEST = `${FORGOT_PASSWORD} FORGOT_PASSWORD_REQUEST`;
-export const FORGOT_PASSWORD_SUCCESS = `${FORGOT_PASSWORD} FORGOT_PASSWORD_SUCCESS`;
-export const FORGOT_PASSWORD_ERROR = `${FORGOT_PASSWORD} FORGOT_PASSWORD_ERROR`;
