+++ src/containers/ForgotPasswordPage/index.js
@@ -2,21 +2,12 @@ import React, { useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { Helmet } from 'react-helmet-async';
 import { useIntl } from 'react-intl';
-import { useInjectReducer } from 'utils/injectReducer';
-import { useInjectSaga } from 'utils/injectSaga';
-import { makeSelectIsForgotPasswordPending } from './selectors';
-import { forgotPassword } from './actions';
+import { makeSelectIsForgotPasswordPending } from 'store/auth/selectors';
+import { forgotPassword } from 'store/auth/actions';
 import ForgotPasswordForm from './ForgotPasswordForm';
-import reducer from './reducer';
-import saga from './saga';
 import messages from './messages';
 
-const key = 'forgotPassword';
-
 function ForgotPasswordPage() {
-  useInjectReducer({ key, reducer });
-  useInjectSaga({ key, saga });
-
   const dispatch = useDispatch();
   const isForgotPasswordPending = useSelector(
     makeSelectIsForgotPasswordPending()
