+++ src/store/auth/actionTypes.js
@@ -9,3 +9,6 @@ export const LOGOUT_REQUEST = `${AUTH} LOGOUT_REQUEST`;
 export const LOGOUT_SUCCESS = `${AUTH} LOGOUT_SUCCESS`;
 export const SET_TOKEN = `${AUTH} SET_TOKEN`;
 export const SESSION_EXPIRED = `${AUTH} SESSION_EXPIRED`;
+export const FORGOT_PASSWORD_REQUEST = `${AUTH} FORGOT_PASSWORD_REQUEST`;
+export const FORGOT_PASSWORD_SUCCESS = `${AUTH} FORGOT_PASSWORD_SUCCESS`;
+export const FORGOT_PASSWORD_ERROR = `${AUTH} FORGOT_PASSWORD_ERROR`;
